+++ CHANGELOG.md
@@ -18,6 +18,8 @@ Changelog for NeoFS Node
 - Use iterators for container listings (#3196)
 - Also reloading shard mode with SIGHUP (#3192)
 - IR logs INFO its state on start based on read FS chain's internals (#3197)
+- Metabase V3->4 migration routine no longer breaks on broken binary or proto-violating object encounter (#3203)
+- Metabase V3->4 migration routine skips objects from the removed containers (#3203)
 
 ### Removed